requirements.txt
crewai
crewai-tools
yfinance
openai
python-dotenv
reportlab
requests